+++ src/iter/mod.rs
@@ -1,24 +1,25 @@
+mod easter;
+mod error;
 mod iterinfo;
+mod masks;
 mod monthinfo;
+mod poslist;
 mod yearinfo;
 
+pub use error::RRuleIterError;
 pub use iterinfo::IterInfo;
-mod poslist;
 pub use poslist::build_poslist;
-mod easter;
-mod masks;
 
 use crate::datetime::{get_weekday_val, DateTime, Time};
 use crate::options::*;
 use crate::utils::{includes, not_empty};
-use chrono::prelude::*;
-use chrono::Duration;
+use chrono::{Datelike, Duration, Timelike};
 
 pub fn decrement_date_until_valid(
     date: DateTime,
     new_month: u32,
     new_year: Option<i32>,
-) -> DateTime {
+) -> Result<DateTime, RRuleIterError> {
     let new_date = if let Some(new_year) = new_year {
         let mut new_date = date.with_year(new_year);
         let mut day_i = 1;
@@ -40,18 +41,21 @@ pub fn decrement_date_until_valid(
             .with_month(new_month);
         day_i += 1;
     }
-    new_date.unwrap()
+    Ok(new_date.unwrap())
 }
 
 pub fn increment_counter_date(
     counter_date: DateTime,
     options: &ParsedOptions,
     filtered: bool,
-) -> DateTime {
+) -> Result<DateTime, RRuleIterError> {
     match options.freq {
-        Frequency::Yearly => counter_date
-            .with_year(counter_date.year() + options.interval as i32)
-            .unwrap(),
+        Frequency::Yearly => {
+            let new_year = counter_date.year() + options.interval as i32;
+            counter_date
+                .with_year(new_year)
+                .ok_or_else(|| RRuleIterError(format!("Year `{}` is invalid", new_year)))
+        }
         Frequency::Monthly => {
             let new_month = counter_date.month() + options.interval as u32;
             if new_month > 12 {
@@ -77,9 +81,9 @@ pub fn increment_counter_date(
             } else {
                 day_delta += -((weekday - options.wkst) as isize) + (options.interval as isize) * 7;
             }
-            counter_date + Duration::days(day_delta as i64)
+            Ok(counter_date + Duration::days(day_delta as i64))
         }
-        Frequency::Daily => counter_date + Duration::days(options.interval as i64),
+        Frequency::Daily => Ok(counter_date + Duration::days(options.interval as i64)),
         Frequency::Hourly => {
             let mut new_hours = counter_date.hour() as usize;
             if filtered {
@@ -98,7 +102,19 @@ pub fn increment_counter_date(
                     break;
                 }
             }
-            counter_date.with_hour(0).unwrap() + Duration::hours(new_hours as i64)
+            // If more then 10 year we assume this is incorrect and will return an error
+            // TODO: This should be checked before iter starts
+            if new_hours > 365 * 24 * 10 {
+                return Err(RRuleIterError(
+                    "Hour interval is to big (more then 10 years), this might be a bug in the settings".to_owned()));
+            }
+            let duration = Duration::hours(new_hours as i64);
+            Ok(counter_date.with_hour(0).ok_or_else(|| {
+                RRuleIterError(format!(
+                    "Could not set hour to `0` for date {}",
+                    counter_date
+                ))
+            })? + duration)
         }
         Frequency::Minutely => {
             let mut minutes_inc = 0;
@@ -124,7 +140,7 @@ pub fn increment_counter_date(
                 }
             }
 
-            counter_date
+            Ok(counter_date)
         }
         Frequency::Secondly => {
             let mut seconds_inc = 0;
@@ -154,41 +170,48 @@ pub fn increment_counter_date(
                 }
             }
 
-            counter_date
+            Ok(counter_date)
         }
     }
 }
 
-pub fn is_filtered(ii: &IterInfo, current_day: usize, options: &ParsedOptions) -> bool {
-    return (!options.bymonth.is_empty() && !options.bymonth.contains(&ii.mmask()[current_day]))
-        || (not_empty(&options.byweekno) && (ii.wnomask().unwrap()[current_day]) == 0)
-        || (not_empty(&options.byweekday)
-            && !includes(&options.byweekday, &(ii.wdaymask()[current_day] as u8)))
-        || (ii.nwdaymask().is_some()
-            && not_empty(ii.nwdaymask().unwrap())
-            && (ii.nwdaymask().unwrap()[current_day]) == 0)
-        || (options.byeaster.is_some()
-            && !(includes(ii.eastermask().unwrap(), &(current_day as isize))))
-        || ((not_empty(&options.bymonthday) || not_empty(&options.bynmonthday))
-            && !includes(&options.bymonthday, &ii.mdaymask()[current_day])
-            && !includes(&options.bynmonthday, &ii.nmdaymask()[current_day]))
-        || (not_empty(&options.byyearday)
-            && ((current_day < ii.yearlen().unwrap()
-                && !includes(&options.byyearday, &(current_day as isize + 1))
-                && !includes(
-                    &options.byyearday.iter().map(|v| *v as isize).collect(),
-                    &(-(ii.yearlen().unwrap() as isize) + current_day as isize),
-                ))
-                || (current_day >= ii.yearlen().unwrap()
-                    && !includes(
-                        &options.byyearday,
-                        &((current_day + 1 - ii.yearlen().unwrap()) as isize),
-                    )
+pub fn is_filtered(
+    ii: &IterInfo,
+    current_day: usize,
+    options: &ParsedOptions,
+) -> Result<bool, RRuleIterError> {
+    // TODO break this up into parts because this is unmaintainable.
+    Ok(
+        (!options.bymonth.is_empty() && !options.bymonth.contains(&ii.mmask()[current_day]))
+            || (not_empty(&options.byweekno) && (ii.wnomask().unwrap()[current_day]) == 0)
+            || (not_empty(&options.byweekday)
+                && !includes(&options.byweekday, &(ii.wdaymask()[current_day] as u8)))
+            || (ii.nwdaymask().is_some()
+                && not_empty(ii.nwdaymask().unwrap())
+                && (ii.nwdaymask().unwrap()[current_day]) == 0)
+            || (options.byeaster.is_some()
+                && !(includes(ii.eastermask().unwrap(), &(current_day as isize))))
+            || ((not_empty(&options.bymonthday) || not_empty(&options.bynmonthday))
+                && !includes(&options.bymonthday, &ii.mdaymask()[current_day])
+                && !includes(&options.bynmonthday, &ii.nmdaymask()[current_day]))
+            || (not_empty(&options.byyearday)
+                && ((current_day < ii.yearlen().unwrap()
+                    && !includes(&options.byyearday, &(current_day as isize + 1))
                     && !includes(
                         &options.byyearday.iter().map(|v| *v as isize).collect(),
-                        &(-(ii.nextyearlen().unwrap() as isize) + current_day as isize
-                            - ii.yearlen().unwrap() as isize),
-                    ))));
+                        &(-(ii.yearlen().unwrap() as isize) + current_day as isize),
+                    ))
+                    || (current_day >= ii.yearlen().unwrap()
+                        && !includes(
+                            &options.byyearday,
+                            &((current_day + 1 - ii.yearlen().unwrap()) as isize),
+                        )
+                        && !includes(
+                            &options.byyearday.iter().map(|v| *v as isize).collect(),
+                            &(-(ii.nextyearlen().unwrap() as isize) + current_day as isize
+                                - ii.yearlen().unwrap() as isize),
+                        )))),
+    )
 }
 
 pub fn remove_filtered_days(
@@ -196,13 +219,13 @@ pub fn remove_filtered_days(
     start: usize,
     end: usize,
     ii: &IterInfo,
-) -> bool {
+) -> Result<bool, RRuleIterError> {
     let mut filtered = false;
 
     for daycounter in start..end {
         match dayset[daycounter] {
             Some(current_day) => {
-                filtered = is_filtered(ii, current_day as usize, &ii.options);
+                filtered = is_filtered(ii, current_day as usize, &ii.options)?;
                 if filtered {
                     dayset[daycounter] = None;
                 }
@@ -210,7 +233,7 @@ pub fn remove_filtered_days(
             None => continue,
         }
     }
-    filtered
+    Ok(filtered)
 }
 
 pub fn build_timeset(options: &ParsedOptions) -> Vec<Time> {
@@ -233,9 +256,13 @@ pub fn build_timeset(options: &ParsedOptions) -> Vec<Time> {
     timeset
 }
 
-pub fn make_timeset(ii: &IterInfo, counter_date: &DateTime, options: &ParsedOptions) -> Vec<Time> {
+pub fn make_timeset(
+    ii: &IterInfo,
+    counter_date: &DateTime,
+    options: &ParsedOptions,
+) -> Result<Vec<Time>, RRuleIterError> {
     if options.freq < Frequency::Hourly {
-        return build_timeset(options);
+        return Ok(build_timeset(options));
     }
 
     if (options.freq >= Frequency::Hourly
@@ -257,7 +284,7 @@ pub fn make_timeset(ii: &IterInfo, counter_date: &DateTime, options: &ParsedOpti
                 .iter()
                 .any(|&s| s == counter_date.second() as usize))
     {
-        return vec![];
+        return Ok(vec![]);
     }
 
     ii.gettimeset(